rename senv_data_parser
