+++ source/producers/ESSStream/CMakeLists.txt
@@ -8,7 +8,7 @@ set(SOURCES
   ${dir}/fb_parser.cpp
   ${dir}/KafkaPlugin.cpp
   ${dir}/mo01_parser.cpp
-  ${dir}/senv_data_parser.cpp
+  ${dir}/SenvParser.cpp
   ${dir}/senv_data_parser_wrong.cpp
   )
 
@@ -20,7 +20,7 @@ set(HEADERS
   ${dir}/fb_parser.h
   ${dir}/KafkaPlugin.h
   ${dir}/mo01_parser.h
-  ${dir}/senv_data_parser.h
+  ${dir}/SenvParser.h
   ${dir}/senv_data_wrong.h
   )
